+++ app/components/developmentToolsComponent/crontabExplainer.tsx
@@ -1,173 +1,24 @@
 "use client";
 import React, { useMemo, useState } from "react";
+import {
+  CRON_ALIASES,
+  CRON_FIELD_SPECS,
+  CronFieldType,
+  humanizeCronExpression,
+  isValidCronExpression,
+  isValidField,
+  parseFieldPartSpec,
+  resolveCronAlias,
+} from "../../libs/cron";
 
 const Cmd = ({ children }: { children: string }) => (
   <code className="px-2 py-1 rounded bg-black/40 border border-white/10 font-mono text-xs">{children}</code>
 );
 
-// Cron aliases mapping
-const ALIASES: Record<string, string> = {
-  "@yearly": "0 0 1 1 *",
-  "@annually": "0 0 1 1 *",
-  "@monthly": "0 0 1 * *",
-  "@weekly": "0 0 * * 0",
-  "@daily": "0 0 * * *",
-  "@midnight": "0 0 * * *",
-  "@hourly": "0 * * * *",
-};
-
-const MONTHS = ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"];
-const DOW = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"];
-type CronFieldType = "minute" | "hour" | "dom" | "month" | "dow";
-
-const FIELD_SPECS: Record<CronFieldType, { min: number; max: number; names?: Record<string, number> }> = {
-  minute: { min: 0, max: 59 },
-  hour: { min: 0, max: 23 },
-  dom: { min: 1, max: 31 },
-  month: {
-    min: 1,
-    max: 12,
-    names: {
-      jan: 1, feb: 2, mar: 3, apr: 4, may: 5, jun: 6,
-      jul: 7, aug: 8, sep: 9, oct: 10, nov: 11, dec: 12,
-    },
-  },
-  dow: {
-    min: 0,
-    max: 6,
-    names: { sun: 0, mon: 1, tue: 2, wed: 3, thu: 4, fri: 5, sat: 6 },
-  },
-};
-
-// Parse a cron field into human-readable text
-const parseField = (field: string, type: "minute" | "hour" | "dom" | "month" | "dow"): string => {
-  if (field === "*") {
-    if (type === "minute") return "every minute";
-    if (type === "hour") return "every hour";
-    if (type === "dom") return "every day";
-    if (type === "month") return "every month";
-    return "every day of week";
-  }
-
-  // Handle step values (*/n or x/n)
-  if (field.includes("/")) {
-    const [base, step] = field.split("/");
-    const stepNum = parseInt(step);
-    if (base === "*") {
-      if (type === "minute") return `every ${stepNum} minute${stepNum > 1 ? "s" : ""}`;
-      if (type === "hour") return `every ${stepNum} hour${stepNum > 1 ? "s" : ""}`;
-      if (type === "dom") return `every ${stepNum} day${stepNum > 1 ? "s" : ""}`;
-      if (type === "month") return `every ${stepNum} month${stepNum > 1 ? "s" : ""}`;
-      return `every ${stepNum} day${stepNum > 1 ? "s" : ""} of week`;
-    }
-    return `every ${stepNum} starting from ${base}`;
-  }
-
-  // Handle ranges (x-y)
-  if (field.includes("-") && !field.includes(",")) {
-    const [start, end] = field.split("-");
-    const startLabel = formatValue(start, type);
-    const endLabel = formatValue(end, type);
-    return `from ${startLabel} to ${endLabel}`;
-  }
-
-  // Handle lists (x,y,z)
-  if (field.includes(",")) {
-    const values = field.split(",").map((v) => formatValue(v.trim(), type));
-    if (values.length > 3) {
-      return `at ${values.slice(0, 3).join(", ")}, and ${values.length - 3} more`;
-    }
-    return `at ${values.join(", ")}`;
-  }
-
-  // Single value
-  return `at ${formatValue(field, type)}`;
-};
-
-const formatValue = (value: string, type: "minute" | "hour" | "dom" | "month" | "dow"): string => {
-  const num = parseInt(value);
-  if (isNaN(num)) {
-    // Handle month/day names
-    if (type === "month") {
-      const monthIndex = MONTHS.findIndex((m) => m.toLowerCase() === value.toLowerCase());
-      return monthIndex >= 0 ? MONTHS[monthIndex] : value;
-    }
-    if (type === "dow") {
-      const dowIndex = DOW.findIndex((d) => d.toLowerCase() === value.toLowerCase());
-      return dowIndex >= 0 ? DOW[dowIndex] : value;
-    }
-    return value;
-  }
-
-  if (type === "month" && num >= 1 && num <= 12) return MONTHS[num - 1];
-  if (type === "dow" && num >= 0 && num <= 6) return DOW[num];
-  if (type === "hour") return `${num.toString().padStart(2, "0")}:00`;
-  if (type === "minute") return `minute ${num}`;
-  return value;
-};
-
-// Generate human-readable explanation
-const explainCron = (cronExpr: string): string => {
-  const trimmed = cronExpr.trim().toLowerCase();
-  
-  // Check for aliases
-  if (ALIASES[trimmed]) {
-    const aliasName = trimmed.substring(1); // Remove @
-    return `Runs ${aliasName} (${ALIASES[trimmed]})`;
-  }
-
-  const parts = cronExpr.trim().split(/\s+/);
-  if (!isValidCronExpression(parts)) {
-    return "Invalid cron expression. Expected 5 fields: minute hour day month weekday";
-  }
-
-  const [minute, hour, dom, month, dow] = parts;
-
-  const minuteText = parseField(minute, "minute");
-  const hourText = parseField(hour, "hour");
-  const domText = parseField(dom, "dom");
-  const monthText = parseField(month, "month");
-  const dowText = parseField(dow, "dow");
-
-  // Build natural sentence
-  let explanation = "Runs ";
-
-  // Time part
-  if (minute === "*" && hour === "*") {
-    explanation += "every minute";
-  } else if (minute !== "*" && hour === "*") {
-    explanation += `${minuteText} of every hour`;
-  } else if (minute === "*" && hour !== "*") {
-    explanation += `every minute ${hourText}`;
-  } else {
-    explanation += `at ${hour.padStart(2, "0")}:${minute.padStart(2, "0")}`;
-  }
-
-  // Day constraints
-  const hasDom = dom !== "*";
-  const hasDow = dow !== "*";
-
-  if (hasDom && hasDow) {
-    explanation += `, ${domText} of the month OR ${dowText}`;
-  } else if (hasDom) {
-    explanation += `, ${domText} of the month`;
-  } else if (hasDow) {
-    explanation += `, ${dowText}`;
-  }
-
-  // Month constraint
-  if (month !== "*") {
-    explanation += `, ${monthText}`;
-  }
-
-  return explanation;
-};
-
 // Calculate next N executions
 const getNextExecutions = (cronExpr: string, count: number = 5): Date[] => {
   try {
-    const trimmed = cronExpr.trim().toLowerCase();
-    const expr = ALIASES[trimmed] || cronExpr.trim();
+    const expr = resolveCronAlias(cronExpr);
     
     const parts = expr.split(/\s+/);
     if (!isValidCronExpression(parts)) return [];
@@ -243,7 +94,7 @@ const formatExecution = (date: Date, timezone: "local" | "utc"): string => {
 };
 
 const getMatchingValues = (field: string, type: CronFieldType): number[] => {
-  const { min, max } = FIELD_SPECS[type];
+  const { min, max } = CRON_FIELD_SPECS[type];
   const values: number[] = [];
   for (let value = min; value <= max; value++) {
     if (matchField(value, field, type)) {
@@ -251,83 +102,6 @@ const getMatchingValues = (field: string, type: CronFieldType): number[] => {
     }
   }
   return values;
-};
-
-const isValidCronExpression = (parts: string[]): boolean => {
-  if (parts.length !== 5) return false;
-  const [minute, hour, dom, month, dow] = parts;
-  return (
-    isValidField(minute, "minute") &&
-    isValidField(hour, "hour") &&
-    isValidField(dom, "dom") &&
-    isValidField(month, "month") &&
-    isValidField(dow, "dow")
-  );
-};
-
-const isValidField = (field: string, type: CronFieldType): boolean => {
-  if (!field || field.trim() === "") return false;
-  const parts = field.split(",");
-  if (parts.some((part) => part.trim() === "")) return false;
-  return parts.every((part) => isValidFieldPart(part.trim(), type));
-};
-
-const isValidFieldPart = (part: string, type: CronFieldType): boolean => {
-  return parseFieldPartSpec(part, type) !== null;
-};
-
-type FieldPartSpec = {
-  start: number;
-  end: number;
-  step: number;
-};
-
-const parseFieldPartSpec = (part: string, type: CronFieldType): FieldPartSpec | null => {
-  const [base, step] = part.split("/");
-  if (!base || part.split("/").length > 2) return null;
-
-  let stepValue = 1;
-  if (step !== undefined) {
-    if (!/^\d+$/.test(step)) return null;
-    stepValue = parseInt(step, 10);
-    if (stepValue <= 0) return null;
-  }
-
-  const spec = FIELD_SPECS[type];
-
-  if (base === "*") {
-    return { start: spec.min, end: spec.max, step: stepValue };
-  }
-
-  if (base.includes("-")) {
-    const [startRaw, endRaw] = base.split("-");
-    if (!startRaw || !endRaw || base.split("-").length > 2) return null;
-    const start = parseFieldValue(startRaw, type);
-    const end = parseFieldValue(endRaw, type);
-    if (start === null || end === null) return null;
-    if (start > end) return null;
-    return { start, end, step: stepValue };
-  }
-
-  const start = parseFieldValue(base, type);
-  if (start === null) return null;
-
-  // x/n means from x through field max, stepping by n.
-  return { start, end: spec.max, step: stepValue };
-};
-
-const parseFieldValue = (raw: string, type: CronFieldType): number | null => {
-  const token = raw.trim().toLowerCase();
-  const spec = FIELD_SPECS[type];
-
-  if (spec.names && token in spec.names) {
-    return spec.names[token];
-  }
-
-  if (!/^\d+$/.test(token)) return null;
-  const value = parseInt(token, 10);
-  if (value < spec.min || value > spec.max) return null;
-  return value;
 };
 
 const matchField = (value: number, field: string, type: CronFieldType): boolean => {
@@ -349,7 +123,7 @@ const CrontabExplainer = () => {
 
   const explanation = useMemo(() => {
     if (!cronInput.trim()) return "";
-    return explainCron(cronInput);
+    return humanizeCronExpression(cronInput);
   }, [cronInput]);
 
   const nextExecutions = useMemo(() => {
@@ -360,7 +134,7 @@ const CrontabExplainer = () => {
   const isValid = useMemo(() => {
     if (!cronInput.trim()) return true;
     const trimmed = cronInput.trim().toLowerCase();
-    if (ALIASES[trimmed]) return true;
+    if (CRON_ALIASES[trimmed]) return true;
     const parts = cronInput.trim().split(/\s+/);
     return isValidCronExpression(parts);
   }, [cronInput]);
